finish excute and abort functions
