+++ Memory CardEmulator/x64/Debug/Memory CardEmulator.log
@@ -1,2 +1,2 @@
-﻿  Memory CardEmulator.cpp
+﻿  Command.cpp
   Memory CardEmulator.vcxproj -> C:\Users\Sabah Tartir\source\repos\Memory CardEmulator\x64\Debug\Memory CardEmulator.exe
